# docs/swagger.yaml
basePath: /
definitions:
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.ConfigRequest:
    properties:
      mtls:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.MTLSConfigRequest'
      oauth2:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.OAuth2ConfigRequest'
      oidc:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.OIDCConfigRequest'
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.CreateAuthRequest:
    properties:
      config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.ConfigRequest'
      enabled:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.MTLSConfigRequest:
    properties:
      allowed_common_names:
        items:
          type: string
        type: array
      allowed_dns_names:
        items:
          type: string
        type: array
      allowed_fingerprints:
        items:
          type: string
        type: array
      ca_cert:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.OAuth2ConfigRequest:
    properties:
      allowed_algorithms:
        items:
          type: string
        type: array
      audiences:
        items:
          type: string
        type: array
      client_id:
        type: string
      client_secret:
        type: string
      introspection_url:
        type: string
      issuer:
        type: string
      jwks_url:
        type: string
      required_scopes:
        items:
          type: string
        type: array
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.OIDCConfigRequest:
    properties:
      allowed_algorithms:
        items:
          type: string
        type: array
      audiences:
        items:
          type: string
        type: array
      issuer:
        type: string
      jwks_url:
        type: string
      public_keys:
        items:
          type: string
        type: array
      required_scopes:
        items:
          type: string
        type: array
      subject_claim:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.UpdateAuthRequest:
    properties:
      config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.ConfigRequest'
      enabled:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.AuthResponse:
    properties:
      api_key:
        description: '#nosec G101'
        type: string
      config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.ConfigResponse'
      created_at:
        type: string
      enabled:
        type: boolean
      gateway_id:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.ConfigResponse:
    properties:
      mtls:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.MTLSConfigResponse'
      oauth2:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.OAuth2ConfigResponse'
      oidc:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.OIDCConfigResponse'
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.ListAuthResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.AuthResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.MTLSConfigResponse:
    properties:
      allowed_common_names:
        items:
          type: string
        type: array
      allowed_dns_names:
        items:
          type: string
        type: array
      allowed_fingerprints:
        items:
          type: string
        type: array
      ca_cert:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.OAuth2ConfigResponse:
    properties:
      allowed_algorithms:
        items:
          type: string
        type: array
      audiences:
        items:
          type: string
        type: array
      client_id:
        type: string
      client_secret:
        type: string
      introspection_url:
        type: string
      issuer:
        type: string
      jwks_url:
        type: string
      required_scopes:
        items:
          type: string
        type: array
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.OIDCConfigResponse:
    properties:
      allowed_algorithms:
        items:
          type: string
        type: array
      audiences:
        items:
          type: string
        type: array
      issuer:
        type: string
      jwks_url:
        type: string
      public_keys:
        items:
          type: string
        type: array
      required_scopes:
        items:
          type: string
        type: array
      subject_claim:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.APIKeyAuthRequest:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      header_name:
        type: string
      header_value:
        type: string
      param_location:
        type: string
      param_name:
        type: string
      param_value:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.EmbeddingConfigRequest:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.APIKeyAuthRequest'
      model:
        type: string
      provider:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.FallbackBudgetRequest:
    properties:
      max_attempts:
        type: integer
      max_total_latency_ms:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.FallbackRequest:
    properties:
      budget:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.FallbackBudgetRequest'
      chain:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      triggers:
        items:
          type: string
        type: array
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.LBConfigRequest:
    properties:
      algorithm:
        type: string
      embedding_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.EmbeddingConfigRequest'
      enabled:
        type: boolean
      members:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.LBPoolMemberRequest'
        type: array
      pool_alias:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.LBPoolMemberRequest:
    properties:
      models:
        items:
          type: string
        type: array
      registry_id:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.ModelPolicyRequest:
    properties:
      allowed:
        items:
          type: string
        type: array
      default:
        type: string
      registry_id:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.ToolkitEntryRequest:
    properties:
      expose_as:
        type: string
      prompt:
        type: string
      registry_id:
        type: string
      resource:
        type: string
      tool:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.UpdateConsumerRequest:
    properties:
      active:
        type: boolean
      fail_mode:
        type: string
      fallback:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.FallbackRequest'
      headers:
        additionalProperties:
          type: string
        type: object
      lb_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.LBConfigRequest'
      model_policies:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.ModelPolicyRequest'
        type: array
      name:
        type: string
      routing_mode:
        type: string
      toolkit:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.ToolkitEntryRequest'
        type: array
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ConsumerResponse:
    properties:
      active:
        type: boolean
      auth_ids:
        items:
          type: string
        type: array
      created_at:
        type: string
      fail_mode:
        type: string
      fallback:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.FallbackResponse'
      gateway_id:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      lb_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.LBConfigResponse'
      model_policies:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ModelPolicyResponse'
        type: array
      name:
        type: string
      registry_ids:
        items:
          type: string
        type: array
      registry_weights:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.RegistryWeightResponse'
        type: array
      role_ids:
        items:
          type: string
        type: array
      routing_mode:
        type: string
      slug:
        type: string
      toolkit:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ToolkitEntryResponse'
        type: array
      type:
        type: string
      updated_at:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.EmbeddingAuthResponse:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      header_name:
        type: string
      header_value:
        type: string
      param_location:
        type: string
      param_name:
        type: string
      param_value:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.EmbeddingConfigResponse:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.EmbeddingAuthResponse'
      model:
        type: string
      provider:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.FallbackBudgetResponse:
    properties:
      max_attempts:
        type: integer
      max_total_latency_ms:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.FallbackResponse:
    properties:
      budget:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.FallbackBudgetResponse'
      chain:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      triggers:
        items:
          type: string
        type: array
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.LBConfigResponse:
    properties:
      algorithm:
        type: string
      embedding_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.EmbeddingConfigResponse'
      enabled:
        type: boolean
      members:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.LBPoolMemberResponse'
        type: array
      pool_alias:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.LBPoolMemberResponse:
    properties:
      models:
        items:
          type: string
        type: array
      registry_id:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ModelPolicyResponse:
    properties:
      allowed:
        items:
          type: string
        type: array
      default:
        type: string
      registry_id:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.RegistryWeightResponse:
    properties:
      registry_id:
        type: string
      weight:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ToolkitEntryResponse:
    properties:
      expose_as:
        type: string
      prompt:
        type: string
      registry_id:
        type: string
      resource:
        type: string
      tool:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    email: support@neuraltrust.ai
    name: NeuralTrust
    url: https://neuraltrust.ai/contact
  description: Administrative API for managing gateways and their registries, policies,
    consumers, roles and auth credentials.
  title: TrustGate Admin API
  version: "1.0"
paths:
...
  /v1/gateways/{gateway_id}/auths:
    get:
      description: Returns a paginated list of auths in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Filter by name (substring match)
        in: query
        name: name
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.ListAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: List auths
      tags:
      - auths
    post:
      consumes:
      - application/json
      description: Creates a new auth in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Auth to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.CreateAuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create an auth
      tags:
      - auths
  /v1/gateways/{gateway_id}/auths/{id}:
    delete:
      description: Deletes an auth from a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Auth id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete an auth
      tags:
      - auths
    get:
      description: Returns a single auth by id.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Auth id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get an auth
      tags:
      - auths
    put:
      consumes:
      - application/json
      description: Updates an existing auth.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Auth id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Auth fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_request.UpdateAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_auth_response.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update an auth
      tags:
      - auths
...
  /v1/gateways/{gateway_id}/consumers/{id}:
    delete:
      description: Deletes a consumer from a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Consumer id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a consumer
      tags:
      - consumers
    get:
      description: Returns a single consumer by id.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Consumer id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ConsumerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a consumer
      tags:
      - consumers
    put:
      consumes:
      - application/json
      description: Updates an existing consumer.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Consumer id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Consumer fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_request.UpdateConsumerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_consumer_response.ConsumerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a consumer
      tags:
      - consumers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"